First based on Status

+++ Contracts/Pagination.cs
@@ -8,14 +8,14 @@ namespace Contracts
 {
     public class Pagination
     {
-        public UInt16 Limit { get; set; }
+        public uint Limit { get; set; }
         public string NextToken { get; set; }
     }
     public class PaginationResponse
     {
         public bool TotalAvailable { get; set; }
-        public UInt64 Total { get; set; }
-        public UInt32 Limit { get; set; }
+        public uint Total { get; set; }
+        public uint Limit { get; set; }
         public string NextToken { get; set; }
     }
